+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/Sec1EcPrivateKeyInfo.kt
@@ -4,20 +4,26 @@
 package at.asitplus.awesn1.crypto
 
 import at.asitplus.awesn1.Asn1BitString
+import at.asitplus.awesn1.Asn1Element
 import at.asitplus.awesn1.Asn1Exception
-import at.asitplus.awesn1.Asn1Integer
+import at.asitplus.awesn1.Asn1OctetString
 import at.asitplus.awesn1.ObjectIdentifier
 import at.asitplus.awesn1.PemLabelSpec
 import at.asitplus.awesn1.WithPemLabel
+import at.asitplus.awesn1.runRethrowing
 import at.asitplus.awesn1.serialization.Asn1Tag
+import at.asitplus.awesn1.serialization.DER
+import at.asitplus.awesn1.serialization.Der
 import at.asitplus.awesn1.serialization.ExplicitlyTagged
+import at.asitplus.awesn1.serialization.decodeFromDer
+import at.asitplus.awesn1.serialization.decodeFromTlv
 import at.asitplus.awesn1.serialization.getValue
-import at.asitplus.awesn1.toInt
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encodeToByteArray
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 /**
@@ -81,8 +87,8 @@ data class Sec1EcPrivateKeyInfo internal constructor(
     override fun hashCode(): Int {
         var result = version.hashCode()
         result = 31 * result + privateKey.contentHashCode()
-        result = 31 * result + (taggedParameters?.hashCode() ?: 0)
-        result = 31 * result + (taggedPublicKey?.hashCode() ?: 0)
+        result = 31 * result + taggedParameters.hashCode()
+        result = 31 * result + taggedPublicKey.hashCode()
         return result
     }
 
@@ -100,6 +106,33 @@ data class Sec1EcPrivateKeyInfo internal constructor(
         const val PEM_LABEL = "EC PRIVATE KEY"
         override val canonicalPemLabel: String get() = PEM_LABEL
 
+        private val EC_PUBLIC_KEY_OID = ObjectIdentifier("1.2.840.10045.2.1")
+
+        operator fun Pkcs8PrivateKeyInfo.Companion.invoke(
+            privateKey: Sec1EcPrivateKeyInfo, curveOid: ObjectIdentifier?,
+            attributes: Set<Asn1Element>? = null, der: Der = DER
+        ) = runRethrowing {
+            Pkcs8PrivateKeyInfo(
+                version = Pkcs8PrivateKeyInfo.Version.V1,
+                privateKeyAlgorithm = X509AlgorithmIdentifier(
+                    EC_PUBLIC_KEY_OID,
+                    curveOid?.encodeToTlv(),
+                ),
+                privateKey = Asn1OctetString(der.encodeToByteArray(privateKey)),
+                attributes = attributes,
+            )
+        }
+
+        fun of(privateKeyInfo: Pkcs8PrivateKeyInfo, der: Der = DER) = runRethrowing {
+            require(privateKeyInfo.algorithmOid == EC_PUBLIC_KEY_OID)
+                { "Pkcs8PrivateKeyInfo is not an RSA private key" }
+            val curveOid = privateKeyInfo.algorithmParameters?.let { der.decodeFromTlv<ObjectIdentifier>(it) }
+            // TODO: is this curve oid somehow relevant?
+            der.decodeFromTlv<Sec1EcPrivateKeyInfo>(
+                privateKeyInfo.privateKey.asEncapsulatingOctetString().element)
+
+        }
+
     }
 
     /**
@@ -116,16 +149,10 @@ data class Sec1EcPrivateKeyInfo internal constructor(
             override val descriptor: SerialDescriptor =
                 PrimitiveSerialDescriptor("Sec1EcPrivateKeyInfo.Version", PrimitiveKind.INT)
 
-            override fun serialize(
-                encoder: Encoder,
-                value: Version
-            ) {
-                encoder.encodeInt(
-                    when(value) {
-                        V1 -> 1
-                    }
-
-                )
+            override fun serialize(encoder: Encoder, value: Version) {
+                encoder.encodeInt(when(value) {
+                    V1 -> 1
+                })
             }
 
             override fun deserialize(decoder: Decoder): Version {